+++ src/data/mockParticipants.ts
@@ -1,15 +1,21 @@
-import { ParticipantDTO } from '../types/participant/participant.type';
+import { UserDTO } from '@/types/user/user.types';
 
-export const mockParticipants: ParticipantDTO['get'][] = [
+export const mockParticipants: UserDTO['get'][] = [
   {
     image:
       'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
+    updatedAt: '2025-04-07T15:00:00',
+    createdAt: '2025-03-02T15:00:00',
     nickname: '종화',
-    updateAt: '2025-04-01T15:00:00',
-    createdAt: '2025-03-26T15:00:00',
+    gender: '남자',
+    email: 'dblls0724@naver.com',
     comment: '방탈출과 스누피를 좋아합니다.',
+    totalGatherings: 3,
+    totalMakeGatherings: 2,
+    totalRE: 8,
+    successCount: 7,
+    failCount: 1,
     representAchievement: [
-      'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
       'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
       'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
     ],
@@ -17,12 +23,18 @@ export const mockParticipants: ParticipantDTO['get'][] = [
   {
     image:
       'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
+    updatedAt: '2025-04-07T15:00:00',
+    createdAt: '2025-03-02T15:00:00',
     nickname: '종화2',
-    updateAt: '2025-04-01T15:00:00',
-    createdAt: '2025-03-26T15:00:00',
+    gender: '남자',
+    email: 'dblls0724@naver.com',
     comment: '방탈출과 스누피를 좋아합니다.',
+    totalGatherings: 3,
+    totalMakeGatherings: 2,
+    totalRE: 8,
+    successCount: 7,
+    failCount: 1,
     representAchievement: [
-      'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
       'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
       'https://i.namu.wiki/i/7gFiOE8l8aznKc3dDwpS1kF1FCBp_eAfGt8xc_CCkRiod_SpDvUdDHXZKltESAjPvLQkSKA-ECTn3zqCSGIZ5g.webp',
     ],
